backend/requirements.txt

google-adk>=0.2.0
google-genai>=0.3.0
google-generativeai>=0.8.0
a2a>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
httpx>=0.27.2
mcp>=1.1.0
psycopg2-binary>=2.9.9
vecs>=0.4.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-exporter-otlp>=1.27.0
structlog>=24.1.0
tenacity>=9.0.0
slowapi>=0.1.9
python-dotenv>=1.0.1
psutil>=5.9.8
